Add support for spell checking roxygen comments
`roxygen2::parse_file()` parses the roxygen comments
in each file. Text from relevant tags is then searched for spelling
errors with `hunspell::hunspell()` to find misspelled words. Because
roxygen does not store the original positions of parsed tags we then
need to find the misspelled word locations in the original roxygen
comment lines of the source. This is done by `find_word_positions()`.